add isolate_from_data and train_from_data, with refactor entrypoint
